ENG-10259: add a login response test
